update main activity
